DiabetesPrediction/backend/requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
joblib
scikit-learn
matplotlib